requirements.txt
Flask>=2.0.0
pymongo>=4.0.0
dnspython>=2.0.0
qrcode>=7.0.0
Pillow>=9.0.0
Werkzeug>=2.0.0
python-dotenv>=1.0.0
colorama>=0.4.6
requests>=2.28.0
langchain>=0.1.0
langchain-community>=0.0.20
langsmith>=0.0.80
PyYAML>=6.0
ollama>=0.1.0
